Update README.md
The last example in the section "Third-pary Field Support" seems to be outdated.

Also there seems to be no way to define object schemas globally and to reuse them. So in the example the Schema `MapPoint` gets recreated for each field and will lead to an error as multiple schemas with the same name exist?